pyproject.toml
[project]
name = "ai-austin-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.5",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
    "streamlit>=1.29.0",
    "langchain-community>=0.0.16",
    "langchainhub>=0.1.14",
    "langchain-chroma>=0.0.1",
    "requests>=2.31.0",
    "pandas>=2.1.3",
    "beautifulsoup4>=4.12.2",
    "openpyxl>=3.1.2",
    "google-generativeai>=0.3.1",
    "pydantic>=2.4.2",
]
